Frontend/patient-portal/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
pymysql
cryptography
python-dotenv
bcrypt
